StorageEventBatchFunction/requirements.txt
azure-functions
azure-batch==6.0.0
azure-storage-blob
